Fixed addback zero degree hits
Zero degree hits were registering as all possible angles in a detector (i.e. 19, 27) when they should all be zero. This was causing them to be placed in the incorrect bins in the 3D histogram making and hence causing issues in the angular correlations. Should be fixed now just have to be careful with normalizations.